MAINT: Move IDBox class to utils.
